requirements.txt
httpx
python-dotenv
httpx-socks[asyncio]
imageio
imageio-ffmpeg
